Add checking for writeability and visualizization
Adds a new directive, remoteWritable, which can be used on
form elements to check if the remote directory is writable.

The test is conducted through creating or opening a new file
in the dirname of the path.  If it is successful, it is assumed
that the file will be writable as well.